Removed dependency on HA build containers

+++ data/include/images.h
@@ -7,7 +7,7 @@ extern "C" const uint8_t _binary_tic_tac_toe_png_start;
 extern "C" const uint8_t _binary_tic_tac_toe_png_end;
 
 uint32_t ticTacToeSize() {
-  return &_binary_tic_tac_toe_png_end - &_binary_tic_tac_toe_png_start;
+  return static_cast<uint32_t>(&_binary_tic_tac_toe_png_end - &_binary_tic_tac_toe_png_start);
 }
 
 #endif
